requirements.txt
antimony
libroadrunner
scipy
python-libsbml
numba
scikit-learn
python-dotenv
pandas 
tqdm
seaborn
matplotlib
joblib
ipykernel
sympy
